refactor(utility): refactor Utility section in  Style
Removed shadow.tsx in a seperate component, same for transparency and borders. split margins and
paddings in subsections in whitespace section
