+++ packages/demo/src/styles/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import {Redirect, Route, RouteComponentProps} from 'react-router-dom';
 
+import Borders from './Borders';
 import Cards from './cards';
 import Colors from './colors';
 import Components from './components';
@@ -9,8 +10,10 @@ import FormControls from './form';
 import GeneralGuidelines from './general-guidelines';
 import Headers from './headers';
 import Icons from './icons';
+import Layout from './Layout';
 import SideMenu from './Menu';
 import Messages from './messages';
+import Shadow from './Shadow';
 import Transparency from './transparency';
 import Typography from './typography';
 import Utility from './utility';
@@ -25,6 +28,7 @@ const Styles: React.FunctionComponent<RouteComponentProps> = ({match}) => {
             <div className="page-content application-container flex flex-column">
                 <div className="wrapper application-main-content m0 flex-auto">
                     <Route path={`${match.url}/general-guidelines`} component={GeneralGuidelines} />
+                    <Route path={`${match.url}/borders`} component={Borders} />
                     <Route path={`${match.url}/cards`} component={Cards} />
                     <Route path={`${match.url}/colors`} component={Colors} />
                     <Route path={`${match.url}/components`} component={Components} />
@@ -32,7 +36,9 @@ const Styles: React.FunctionComponent<RouteComponentProps> = ({match}) => {
                     <Route path={`${match.url}/form-controls`} component={FormControls} />
                     <Route path={`${match.url}/headers`} component={Headers} />
                     <Route path={`${match.url}/icons`} component={Icons} />
+                    <Route path={`${match.url}/layout`} component={Layout} />
                     <Route path={`${match.url}/messages`} component={Messages} />
+                    <Route path={`${match.url}/shadow`} component={Shadow} />
                     <Route path={`${match.url}/transparency`} component={Transparency} />
                     <Route path={`${match.url}/typography`} component={Typography} />
                     <Route path={`${match.url}/utility`} component={Utility} />
